Updated 3rd party webrtc code
// FREEBIE

+++ jni/webrtc/common_audio/signal_processing/cross_correlation_mips.c
@@ -13,10 +13,10 @@
 void WebRtcSpl_CrossCorrelation_mips(int32_t* cross_correlation,
                                      const int16_t* seq1,
                                      const int16_t* seq2,
-                                     int16_t dim_seq,
-                                     int16_t dim_cross_correlation,
-                                     int16_t right_shifts,
-                                     int16_t step_seq2) {
+                                     size_t dim_seq,
+                                     size_t dim_cross_correlation,
+                                     int right_shifts,
+                                     int step_seq2) {
 
   int32_t t0 = 0, t1 = 0, t2 = 0, t3 = 0, sum = 0;
   int16_t *pseq2 = NULL;
