+++ jni/webrtc/common_audio/signal_processing/resample_by_2_mips.c
@@ -29,11 +29,11 @@ static const uint16_t kResampleAllpass2[3] = {12199, 37471, 60255};
 
 // decimator
 void WebRtcSpl_DownsampleBy2(const int16_t* in,
-                             int16_t len,
+                             size_t len,
                              int16_t* out,
                              int32_t* filtState) {
   int32_t out32;
-  int16_t i, len1;
+  size_t i, len1;
 
   register int32_t state0 = filtState[0];
   register int32_t state1 = filtState[1];
