+++ jni/webrtc/modules/audio_coding/codecs/ilbc/cb_update_best_index.h
@@ -24,13 +24,13 @@
 void WebRtcIlbcfix_CbUpdateBestIndex(
     int32_t CritNew,    /* (i) New Potentially best Criteria */
     int16_t CritNewSh,   /* (i) Shift value of above Criteria */
-    int16_t IndexNew,   /* (i) Index of new Criteria */
+    size_t IndexNew,   /* (i) Index of new Criteria */
     int32_t cDotNew,    /* (i) Cross dot of new index */
     int16_t invEnergyNew,  /* (i) Inversed energy new index */
     int16_t energyShiftNew,  /* (i) Energy shifts of new index */
     int32_t *CritMax,   /* (i/o) Maximum Criteria (so far) */
     int16_t *shTotMax,   /* (i/o) Shifts of maximum criteria */
-    int16_t *bestIndex,   /* (i/o) Index that corresponds to
+    size_t *bestIndex,   /* (i/o) Index that corresponds to
                                    maximum criteria */
     int16_t *bestGain);   /* (i/o) Gain in Q14 that corresponds
                                    to maximum criteria */
