+++ jni/webrtc/modules/audio_coding/codecs/isac/fix/source/arith_routins.h
@@ -74,7 +74,7 @@ int16_t WebRtcIsacfix_EncTerminate(Bitstr_enc *streamData);
  * Return value             : number of bytes in the stream so far
  *                            <0 if error detected
  */
-int16_t WebRtcIsacfix_DecLogisticMulti2(
+int WebRtcIsacfix_DecLogisticMulti2(
     int16_t *data,
     Bitstr_dec *streamData,
     const int32_t *env,
