+++ jni/webrtc/modules/audio_coding/neteq/test/NETEQTEST_RTPpacket.h
@@ -14,7 +14,7 @@
 #include <map>
 #include <stdio.h>
 #include "webrtc/typedefs.h"
-#include "webrtc/modules/interface/module_common_types.h"
+#include "webrtc/modules/include/module_common_types.h"
 
 enum stereoModes {
     stereoModeMono,
@@ -36,13 +36,12 @@ public:
     int readFixedFromFile(FILE *fp, size_t len);
     virtual int writeToFile(FILE *fp);
     void blockPT(uint8_t pt);
-    //int16_t payloadType();
     virtual void parseHeader();
     void parseHeader(webrtc::WebRtcRTPHeader* rtp_header);
     const webrtc::WebRtcRTPHeader* RTPinfo() const;
     uint8_t * datagram() const;
     uint8_t * payload() const;
-    int16_t payloadLen();
+    size_t payloadLen();
     int16_t dataLen() const;
     bool isParsed() const;
     bool isLost() const;
@@ -73,7 +72,7 @@ public:
     uint8_t *       _payloadPtr;
     int                 _memSize;
     int16_t         _datagramLen;
-    int16_t         _payloadLen;
+    size_t          _payloadLen;
     webrtc::WebRtcRTPHeader _rtpInfo;
     bool                _rtpParsed;
     uint32_t        _receiveTime;
